fix(booking): change type of vehicleId from Integer to Long

+++ src/main/java/com/swp/autocarwash/booking/event/BookingCanceledEvent.java
@@ -10,7 +10,7 @@ import java.time.Instant;
 public class BookingCanceledEvent {
 
     private final Long bookingId;
-    private final Integer vehicleId;
+    private final Long vehicleId;
     private final Long customerId;
     private final Long canceledByStaffId;
     private final String bookingType;
